+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - rename `DB_URI` to `DATABASE_URI`
 - rename the `db_uri` argument of `init_sqlalchemy_unchained` to `database_uri`
 - remove the `ModelMetaOptionsFactory._model_repr` property
+- remove the `validates` decorator
 
 ## v0.5.1 (2018/10/20)